--- Ambulance Zorg/Assets/BodyComponentBehavior.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 326f94eaad77cf64688c45bf228473e2
-timeCreated: 1476168935
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
