--- Ambulance Zorg/Assets/CrosshairBehavior.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 18d44406a65926f43859f0dd57032325
-timeCreated: 1476169445
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
