+++ Ambulance Zorg/Assets/GameControllerBehavior.cs
@@ -14,8 +14,6 @@ public class GameControllerBehavior : MonoBehaviour {
 
     // Use this for initialization
     void Start () {
-
-		Cursor.lockState = CursorLockMode.Locked;
         
 
     }
